开机场景，应用连续启动场景识别

+++ cgroup_sched/framework/sched_controller/include/app_startup_scene_rec.h
@@ -37,7 +37,7 @@ private:
     void CleanRecordSceneData();
     void UpdateAppStartupNum(std::string uid, int64_t curTime, std::string bundleName);
     bool IsContinuousStartup();
-    int64_t appStartCount_ = 0;
+    int32_t appStartCount_ = 0;
     int64_t lastAppStartTime_ = 0;
     std::atomic_bool isReportContinuousStartup_ = false;
     std::string lastStartUid_ = "";
